Allergen CRUD

+++ src/main/java/com/cooker/cook/repositories/AllergenRepository.java
@@ -3,5 +3,5 @@ package com.cooker.cook.repositories;
 import com.cooker.cook.entities.Allergen;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AllergenRepository extends JpaRepository <Allergen,Integer> {
+public interface AllergenRepository extends JpaRepository <Allergen,Long> {
 }
